CRUD stuff is working now. Need to integrate React Router

+++ models/Api/ThreeJSViewer.php
@@ -19,23 +19,30 @@ class Api_ThreeJSViewer extends Omeka_Record_Api_AbstractRecordAdapter
         $record->enable_shaders = $data->enable_shaders;
         $record->enable_measurement = $data->enable_measurement;
         $record->model_units = $data->model_units;
+        $record->needs_delete = $data->needs_delete;
     }
 
     // Set data to a record during a PUT request.
     public function setPutData(Omeka_Record_AbstractRecord $record, $data)
     {
-        // Set properties directly to an existing record.
-        if ($data->three_file_id) {
-          $current = get_record_by_id('File', $record->three_file_id);
-          $record->three_file_id = $data->three_file_id;
-          $current->delete();
+        if ($data->needs_delete) {
+          // Don't care about setting anything else as this will be deleted on the next after_save_record hook
+          $record->needs_delete = $data->needs_delete;
+        } else {
+          // Set properties directly to an existing record.
+          if ($data->three_file_id) {
+            $current = get_record_by_id('File', $record->three_file_id);
+            $record->three_file_id = $data->three_file_id;
+            $current->delete();
+          }
+          $record->item_id = $data->item_id;
+          $record->background_url = $data->background_url;
+          $record->enable_lights = $data->enable_lights;
+          $record->enable_materials = $data->enable_materials;
+          $record->enable_shaders = $data->enable_shaders;
+          $record->enable_measurement = $data->enable_measurement;
+          $record->model_units = $data->model_units;
+          $record->needs_delete = $data->needs_delete;
         }
-        $record->item_id = $data->item_id;
-        $record->background_url = $data->background_url;
-        $record->enable_lights = $data->enable_lights;
-        $record->enable_materials = $data->enable_materials;
-        $record->enable_shaders = $data->enable_shaders;
-        $record->enable_measurement = $data->enable_measurement;
-        $record->model_units = $data->model_units;
     }
 }
